refactoring. creating new routes
