+++ rest-server/controllers/v1/index.js
@@ -1,5 +1,7 @@
+const seasonController = require("./seasonController");
 const userController = require("./userController");
 
 module.exports = {
-  userController
+  seasonController,
+  userController,
 };
