+++ rest-server/routes/v1/index.js
@@ -3,8 +3,10 @@ const router = express.Router();
 
 // Import v1 user routes
 const userRoutes = require("./userRoutes");
+const seasonRoutes = require("./seasonRoutes");
 
 // Mount v1 user routes under /v1/user
 router.use("/user", userRoutes);
+router.use("/season", seasonRoutes);
 
 module.exports = router;
